Move source files into correct relative dirs
Needed so that Merlin in the original project folder (ie. when
DUNE_BUILD_DIR is set as <project>/_build) can locate the source
code. After a dune build Dune has embedded in the _build\default\.merlin-conf\exe-QuickstartMain
(etc.) files the location of the source code used for compilation.

And rename DUNE_BUILD_DIR to DUNE_PROJECT_DIR so can know when we have to shift the relative path.