Correction: Do not propagate DKML_HOST_ABI
It is hardcoded into DkRun bytecode executables that are downloaded by ./dk (so ABI hardcoding comes from ./dk).

But not good to have two sources of truth ... in the future DkRun may have a better detection of ABI (ex. ./dk downloads x86_64 for macOS but ABI is detected as arm64).